+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## 1.1.0
+November 12, 2023
+- InheritedWidget State class no public, InheritedBoxWidgetState
+- InheritedBoxWidgetState.notified used in updateShouldNotify() function
+
 ## 1.0.0
 November 09, 2023
 - Initial commit